Make "erase data" procedure two-step
