+++ core-bundle/src/Controller/Backend/EraseDataController.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Ferienpass package.
+ *
+ * (c) Richard Henkenjohann <richard@ferienpass.online>
+ *
+ * For more information visit the project website <https://ferienpass.online>
+ * or the documentation under <https://docs.ferienpass.online>.
+ */
+
+namespace Ferienpass\CoreBundle\Controller\Backend;
+
+use Ferienpass\CoreBundle\Facade\EraseDataFacade;
+use Ferienpass\CoreBundle\Form\SimpleType\ContaoRequestTokenType;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+/**
+ * @Route("/erase-data", name="backend_erase_data")
+ */
+final class EraseDataController extends AbstractBackendController
+{
+    private EraseDataFacade $eraseDataFacade;
+
+    public function __construct(EraseDataFacade $eraseDataFacade)
+    {
+        $this->eraseDataFacade = $eraseDataFacade;
+    }
+
+    public function __invoke(Request $request): Response
+    {
+        $this->initializeContaoFramework();
+
+        $participants = $this->eraseDataFacade->expiredParticipants();
+        if (empty($participants)) {
+            return $this->redirectToRoute('contao_backend');
+        }
+
+        $form = $this->createFormBuilder()->add('requestToken', ContaoRequestTokenType::class)->getForm();
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->eraseDataFacade->eraseData();
+
+            return $this->redirectToRoute('contao_backend');
+        }
+
+        return $this->render('@FerienpassCore/Backend/erase-data.html.twig', [
+            'participants' => $participants,
+            'form' => $form->createView(),
+        ]);
+    }
+}
