+++ core-bundle/templates/Backend/Dashboard/erase_data.html.twig
@@ -1,21 +1,26 @@
-<div class="bg-white shadow sm:rounded-lg">
+<div class="bg-white shadow sm:rounded-lg max-w-2xl my-8">
     <div class="px-4 py-5 sm:p-6">
-        <h3 class="text-lg leading-6 font-medium text-gray-900">
-            Personenbezogene Daten
+        <h3 class="text-lg leading-6 font-medium text-gray-900 flex items-center">
+            <span class="flex-shrink-0">
+                <span class="block h-5 w-5 text-red-400 mr-1" aria-hidden="true">
+                    {{ source('@FerienpassCore/Icons/shield-exclamation-solid.svg') }}
+                </span>
+            </span>
+                <span>Personenbezogene Daten</span>
         </h3>
         <div class="mt-2 max-w-xl text-sm leading-5 text-gray-500">
-            <p>
-                Personenbezogene Daten von {{ count }} Kindern (+ Eltern) sind zum Löschen bereit. Bitte bestätigen Sie
-                den Vorgang.
-            </p>
+            <p>Aus Datenschutzgründen werden alle personenbezogenen Daten von Kindern aus vergangenen Veranstaltungen
+                gelöscht.</p>
         </div>
-        <div class="mt-5">
-            {{ form_start(form) }}
-            <button type="submit"
-                    class="inline-flex items-center justify-center px-4 py-2 border border-transparent font-medium rounded-md text-red-700 bg-red-100 hover:bg-red-50 focus:outline-none focus:border-red-300 focus:shadow-outline-red active:bg-red-200 transition ease-in-out duration-150 sm:text-sm sm:leading-5">
-                Daten bereinigen
-            </button>
-            {{ form_end(form) }}
+        <div class="mt-2 max-w-xl text-sm leading-5 text-gray-700">
+            <p class="font-medium">Die Daten von {{ participants|length }} Kindern (+ Eltern) müssen aus
+                dem System gelöscht werden.</p>
+        </div>
+        <div class="mt-3 text-sm leading-5">
+            <a href="{{ path('backend_erase_data') }}"
+               class="font-medium text-primary-600 hover:text-primary-500 transition ease-in-out duration-150">
+                Zum Löschen &rarr;
+            </a>
         </div>
     </div>
 </div>
