+++ core-bundle/templates/Backend/erase-data.html.twig
@@ -1,0 +1,65 @@
+{% extends "@ContaoCore/Backend/be_page.html.twig" %}
+
+{% block headline %}
+    Vorherige personenbezogene Daten löschen
+{% endblock %}
+
+{% block error %}
+    {%- if error is defined -%}
+        {{- error -}}
+    {%- endif -%}
+{% endblock %}
+
+{% block main %}
+    <div class="bg-white shadow sm:rounded-lg">
+        <div class="px-4 py-5 sm:p-6">
+            <h3 class="text-lg leading-6 font-medium text-gray-900">
+                Daten sofort löschen
+            </h3>
+            <div class="mt-2 max-w-xl text-sm leading-5 text-gray-500">
+                <p>
+                    Personenbezogene Daten untenstehenden Kindern ({{ participants|length }} Kinder + Eltern) sind zum
+                    Löschen bereit. Bitte bestätigen Sie den Vorgang.
+                </p>
+            </div>
+            <div class="mt-5">
+                {{ form_start(form) }}
+                <button type="submit"
+                        class="inline-flex items-center justify-center px-4 py-2 border border-transparent font-medium rounded-md text-red-700 bg-red-100 hover:bg-red-50 focus:outline-none focus:border-red-300 focus:shadow-outline-red active:bg-red-200 transition ease-in-out duration-150 sm:text-sm sm:leading-5">
+                    Daten bereinigen
+                </button>
+                {{ form_end(form) }}
+            </div>
+        </div>
+    </div>
+
+    <div class="mt-8">
+        <table class="w-full">
+            <thead>
+            <tr>
+                <th class="px-4 py-3 border-b border-gray-200 bg-gray-50 text-left text-sm leading-4 font-medium text-gray-500">
+                    Name
+                </th>
+                <th class="px-4 py-3 border-b border-gray-200 bg-gray-50 text-left text-sm leading-4 font-medium text-gray-500">
+                    letzte Anmeldung
+                </th>
+            </tr>
+            </thead>
+            <tbody>
+            {% for participant in participants %}
+                <tr class="border-t border-gray-200">
+                    <td class="px-4 py-2">
+                        <p class="text-sm leading-5 font-medium text-gray-500">
+                            {{ participant.firstname  ~ ' ' ~ participant.lastname }}
+                        </p>
+                    </td>
+                    <td class="px-4 py-2">
+                        {% set attendance = participant.lastAttendance %}
+                        {{ (attendance and attendance.offer.edition) ? attendance.offer.edition.name : '-' }}
+                    </td>
+                </tr>
+            {% endfor %}
+            </tbody>
+        </table>
+    </div>
+{% endblock %}
